requirements.txt
python-telegram-bot==13.11
pdf2docx==0.5.6
Pillow==8.2.0
tqdm==4.61.1
